Grocery: quantity accepts decimals; USER_GUIDE login + grocery updates
Switch grocery_items.quantity from Integer to Numeric(10, 3) so
fractional amounts like 1.5 lb can be expressed without stuffing the
decimal into the unit field. Updates the form parser, the LLM tool
schema (GroceryItemArgs.quantity), and adds a no-op alembic migration
that alters the column in place.

USER_GUIDE: rewrite Logging in (two-user model, 30-day session, link to
OPERATIONS for hash reset) and rewrite Grocery (decimal quantities,
unit is free text, explicit note that the list isn't deduplicated).

+++ src/family_assistant/ai_gateway/tools.py
@@ -55,7 +55,7 @@ class _StrictModel(BaseModel):
 class GroceryItemArgs(_StrictModel):
     name: str
     category: str | None = None
-    quantity: int | None = None
+    quantity: Decimal | None = None
     unit: str | None = None
     notes: str | None = None
